drivers/qma6100p: fix acceleration data truncation
µg values from qma6100p_read overflowed the int16_t fields, turning
gravity into garbage.

- Store µg in int32_t in qma6100p_data_t.
- In SAUL, convert µg to mg (/1000), scale -3. Worst case
8192 * 3810 / 1000 = 31211, fits int16_t.

+++ drivers/include/qma6100p.h
@@ -216,9 +216,9 @@ typedef struct {
  * @brief   Converted acceleration data in g
  */
 typedef struct {
-    int16_t x; /**< acceleration in X direction [ug] */
-    int16_t y; /**< acceleration in Y direction [ug] */
-    int16_t z; /**< acceleration in Z direction [ug] */
+    int32_t x; /**< acceleration in X direction [ug] */
+    int32_t y; /**< acceleration in Y direction [ug] */
+    int32_t z; /**< acceleration in Z direction [ug] */
 } qma6100p_data_t;
 
 /**
